Allow switching from Intel without reboot
Switching from Intel to NVIDIA or Hybrid, and between NVIDIA and Hybrid,
can easily be accomplished without the need to reboot. Users will only
have to log out and log in for the changes to take effect.

Switching from NVIDIA or Hybrid *to* Intel without a reboot is not as
trivial, and not included in this change. It requires a restart of GDM,
completely interrupting the user's session. It can still be done with:

    system76-power graphics intel
    systemctl restart display-manager

See: #57